Tentando realizar a integração com SOAP.

+++ app/src/main/java/com/example/tccapp/list/TeachersItemViewModel.kt
@@ -8,6 +8,6 @@ class TeachersItemViewModel(
     val itemAdapterOnClickEvent: SingleLiveEvent<Int>
 ) {
     fun onClick() {
-        itemAdapterOnClickEvent.postValue(item.id)
+        itemAdapterOnClickEvent.postValue(item.id.toInt())
     }
 }
